requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.1